requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.59.0
django-autoslug==1.9.9
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-dirtyfields==1.9.2
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-phonenumber-field==7.3.0
django-resized==1.0.2
idna==3.6
mysqlclient==2.2.1
oauthlib==3.2.2
phonenumberslite==8.13.27
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
uuid==1.30
